More keybindings added to standalone viewer, failsafe when trying to delete elements when there are no more

+++ caseron/view/gui/integrableviewer.py
@@ -211,6 +211,9 @@ class IntegrableViewer(QOpenGLWidget):
         self.update()
 
     def delete(self, _id: int) -> None:
+        if _id < 0:
+            return
+
         self.makeCurrent()
         self.model.delete(_id)
         self.drawable_collection.delete(_id)
